tests/wpt/web-platform-tests/scroll-animations/css/scroll-timeline-sampling.html

<!DOCTYPE html>
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#avoiding-cycles">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<style>
  #scroller {
    overflow: hidden;
    width: 100px;
    height: 100px;
    scroll-timeline: timeline;
  }
  #contents {
    height: 200px;
  }
  @keyframes expand {
    from { width: 100px; }
    to { width: 200px; }
  }
  #element {
    width: 0px;
    animation: expand 10s linear;
    animation-timeline: timeline;
  }
  /* Ensure stable expectations if feature is not supported */
  @supports not (animation-timeline:foo) {
    #element { animation-play-state: paused; }
  }
</style>
<div id=scroller>
  <div id=contents></div>
</div>
<div id=element></div>
<script>
  promise_test(async (t) => {
    // The scroll timeline is initially inactive until the first frame.
    assert_equals(getComputedStyle(element).width, '0px');
    await waitForNextFrame();
    scroller.scrollTop = 50;
    // Scrolling position should not yet be reflected in the animation,
    // since the new scroll position has not yet been sampled.
    assert_equals(getComputedStyle(element).width, '100px');
    await waitForNextFrame();
    assert_equals(getComputedStyle(element).width, '150px');
  }, 'Scroll position is sampled once per frame');
</script>
