LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-sampling-expected.txt


FAIL Scroll position is sampled once per frame assert_equals: expected "150px" but got "100.15625px"
